rearrange existing tests into new file structure
