+++ tests/test_bgc.py
@@ -13,42 +13,6 @@ from ._helpers import (
 )
 
 
-def test_bgc_real_bec(tmp_path, input_dir, reference_results):
-    cpp_keys = (
-        UNIVERSAL_CPP_KEYS + OCEAN_PHYSICS_CPP_KEYS
-        + REALISTIC_CPP_KEYS + BGC_REALISTIC_CPP_KEYS
-        + ["TIDES", "BIOLOGY_BEC2"]
-    )
-    conf = ROMSConfiguration(cpp_keys=cpp_keys, location=tmp_path / "build")
-    conf.compile()
-
-    nml = create_test_namelist_dict(input_dir)
-    nml["TIME_STEPPING"]["dt"] = 20
-    nml["TIME_STEPPING"]["ntimes"] = 10
-    nml["BASIC_OUTPUT_SETTINGS"]["output_period_his"] = 100
-    nml["BGC_SETTINGS"]["output_period_his"] = 400
-    nml["PARAM_SETTINGS"]["ntrc_bio"] = 26
-    nml["TIDES_SETTINGS"].update({"pot_tides": True, "ntides": 2})
-    nml["FORCING_FILES"] = {
-        "frcfile": [
-            str(input_dir / "example_input_boundary_forcing.nc"),
-            str(input_dir / "example_input_surface_forcing.nc"),
-            str(input_dir / "example_input_bgc_surface_forcing.nc"),
-            str(input_dir / "example_input_river_forcing.nc"),
-            str(input_dir / "example_input_bgc_boundary_forcing.nc"),
-            str(input_dir / "example_input_tides.nc"),
-        ]
-    }
-
-    conf.run(nml)
-
-    pv = get_summary_value(conf.location, prefix="roms_his.20100101000000")
-    bv = get_summary_value(conf.location, prefix="roms_bgc.20100101000000")
-    assert_output_matches_reference(
-        reference_results, "bgc_real_bec", {"physics": pv, "bgc": bv}
-    )
-
-
 def test_bgc_real_marbl(tmp_path, input_dir, reference_results):
     cpp_keys = (
         UNIVERSAL_CPP_KEYS + OCEAN_PHYSICS_CPP_KEYS
